multimeeks/requirements.txt
asgiref              3.8.1
async-timeout        4.0.3
attrs                23.2.0
autobahn             23.6.2
Automat              22.10.0
certifi              2024.2.2
cffi                 1.16.0
channels             4.1.0
channels-redis       4.2.0
constantly           23.10.4
coverage             5.5
cryptography         42.0.7
daphne               4.1.2
Django               5.0.6
django-countries     7.6.1
django-tastypie      0.14.7
django-widget-tweaks 1.5.0
djangorestframework  3.15.1
h11                  0.14.0
hyperlink            21.0.0
idna                 3.7
incremental          22.10.0
msgpack              1.0.8
mysqlclient          2.2.4
outcome              1.3.0.post0
pillow               10.3.0
pip                  23.0.1
pyasn1               0.6.0
pyasn1_modules       0.4.0
pycparser            2.22
pyOpenSSL            24.1.0
PySocks              1.7.1
python-dateutil      2.9.0.post0
python-mimeparse     1.6.0
redis                5.0.4
selenium             4.21.0
service-identity     24.1.0
setuptools           66.1.1
six                  1.16.0
sniffio              1.3.1
sortedcontainers     2.4.0
sqlparse             0.5.0
trio                 0.25.1
trio-websocket       0.11.1
Twisted              24.3.0
txaio                23.1.1
typing_extensions    4.11.0
urllib3              2.2.1
wsproto              1.2.0
zope.interface       6.4